First-edition of Unlock Pi-script, using webservice
This first version has a few obvious issues:

1) Constant hostname for webservice
2) Not sure if this kind of calling system is the one we
   want to use
3) Obviously, the console.log output is not going to look like
   that forever. It's just there to make debugging easier without
   a proper test unit.

The idea is that the RFID reader should trigger this script to
execute when a card is swiped, and it should pass 'type=rfid' and
'id=<the parsed ID of the RFID card>', which will then get passed
off to the webservice who decides whether or not we want to unlock
the door. So most of the hard work happens on the web service.

As far as automated testing is concerned... We can certainly test
that the client gets an expected response from the web service on
a regular basis. But I think I will be writing this test in a
subsequent patch.