Add Logging functionality to webservice
Logging is implemented in a kind of stupid way, the data model isn't
terribly intuitive, but I THINK this all basically works:

An event generated by this for an authorized access looks something
like this:

```javascript
{
	"device_id" : ObjectId("51e562d2eb3e02a33c28efef"),
	"rfid_id" : ObjectId("51e562d2eb3e02a33c28eff2"),
	"entry_time" : ISODate("2013-07-16T17:39:15.463Z"),
	"_id" : ObjectId("51e585437fd6387754000003"),
	"status" : "authorized",
	"__v" : 0
}
```

If you notice, I also have some other status strings that I use,
depending on the situation.

Obviously user info can be inferred from an RFID, but it might be
a good idea to store the user that was assigned to that RFID at
the time of the access, if we are in fact going to support reassignment
of RFIDS.

Lets talk about it.