Reports UI V1 + Data model changes
Some basic stuff has been added for reports generation

The returned dataset isn't sorted or anything, there's no pagination,
(I'm probably not going to bother with pagination, the dataset can't
possibly get that big, and it feels really awkward), but it does
display an actual dataset.

The Event object's data has been changed a fair bit:

- Store username, RFID tag, device name AND hostname as Strings,
rather than using references to lookup separate subdocuments.
This makes querying a lot simpler, and I think in the case of
events it probably makes them smaller, byte-for-byte (every ObjectId
is 24bytes in string form, 48bytes in most browsers (utf16)), so you
multiply that by 2 and you get nearly 50 bytes, which is larger than a
typical Event entry without a user's full name, and certainly without a
Lock name. But if we do a fair comparison and include the Username, that
makes the ObjectId version over a dozen bytes large on average! So I
don't really expect anyone to get seriously damaged by this, so long as
very bad crap is kept out of the database, amirite?

- I've renamed entry_time to entryTime, because everything else seems to
be camelcased, so consistency is key and all that.

- I've updated mongo_setup, and have tested it, it works beautifully.

- I have updated the webservice, but have not tested it yet -- I will
make sure that it works mostly correclty tomorrow, and will also write
some mochatests for the events.findWithParams routine. Mostly stuff is
working, though, woo.