Implement Search/Filtering in Reports UI
There are a lot of regexes here to simulate (sort of) SQL LIKE
expressions. This kind of sucks and all, case-insensitive is
kind of desirable though and case insensitive means that can't
really use the optimized stuff designed for indexed fields, so
it's lose-lose no matter what.

But whatever, the data set is small and probably always will be,
and it's not really slow or anything.