Events/Reports Chrome Bugs, recent work
As I said on Skype, for whatever reason, Chrome does not seem to
send the $resource XHR requests for reports. The initial query is
executed, but subsequent ones never seem to go through.

I know this had been working before, and I'm not sure what broke.

Please take a look and see if you can fix it ;-;